+++ storage/framework/core/storage/src/delete.ts
@@ -1,11 +1,11 @@
-import type { Result } from '@stacksjs/error-handling'
-import { join } from '@stacksjs/path'
-import { err, handleError, ok } from '@stacksjs/error-handling'
 import { italic, log } from '@stacksjs/cli'
+import type { Result } from '@stacksjs/error-handling'
+import { err, handleError, ok } from '@stacksjs/error-handling'
+import { join } from '@stacksjs/path'
 import { isFile } from './files'
 import { isFolder } from './folders'
-import { glob } from './glob'
 import { fs } from './fs'
+import { glob } from './glob'
 
 export function deleteFolder(path: string): Promise<Result<string, Error>> {
   return new Promise((resolve, reject) => {
@@ -16,54 +16,55 @@ export function deleteFolder(path: string): Promise<Result<string, Error>> {
       }
 
       return resolve(ok(`Path ${path} was not a directory`))
-    }
-    catch (error) {
+    } catch (error) {
       return reject(err(error))
     }
   })
 }
 
-export async function isDirectoryEmpty(path: string): Promise<Result<boolean, Error>> {
+export async function isDirectoryEmpty(
+  path: string,
+): Promise<Result<boolean, Error>> {
   return new Promise((resolve, reject) => {
     try {
       if (fs.statSync(path).isDirectory()) {
-        if (fs.readdirSync(path).length === 0)
-          return resolve(ok(true))
-        else
-          return resolve(ok(false))
+        if (fs.readdirSync(path).length === 0) return resolve(ok(true))
+        else return resolve(ok(false))
       }
 
       return resolve(ok(false))
-    }
-    catch (error) {
+    } catch (error) {
       return reject(err(error))
     }
   })
 }
 
-export async function deleteEmptyFolder(path: string): Promise<Result<string, Error>> {
+export async function deleteEmptyFolder(
+  path: string,
+): Promise<Result<string, Error>> {
   return new Promise((resolve, reject) => {
     try {
       if (fs.statSync(path).isDirectory()) {
         if (fs.readdirSync(path).length === 0) {
           fs.rmSync(path, { recursive: true, force: true })
           return resolve(ok(`Deleted ${path}`))
+        } else {
+          return resolve(ok(`Path ${path} was not empty`))
         }
-        else { return resolve(ok(`Path ${path} was not empty`)) }
       }
 
       return resolve(ok(`Path ${path} was not a directory`))
-    }
-    catch (error) {
+    } catch (error) {
       return reject(err(error))
     }
   })
 }
 
-export async function deleteEmptyFolders(dir: string): Promise<Result<string, Error>> {
+export async function deleteEmptyFolders(
+  dir: string,
+): Promise<Result<string, Error>> {
   try {
-    if (!fs.existsSync(dir))
-      return ok(`Path ${dir} does not exist`)
+    if (!fs.existsSync(dir)) return ok(`Path ${dir} does not exist`)
 
     const files = fs.readdirSync(dir)
     for (const file of files) {
@@ -71,14 +72,12 @@ export async function deleteEmptyFolders(dir: string): Promise<Result<string, Er
       if (isFolder(p)) {
         if (fs.readdirSync(p).length === 0)
           fs.rmSync(p, { recursive: true, force: true })
-        else
-          await deleteEmptyFolders(p)
+        else await deleteEmptyFolders(p)
       }
     }
 
     return ok(`Deleted empty folders located in ${dir}`)
-  }
-  catch (error: any) {
+  } catch (error: any) {
     return err(error)
   }
 }
@@ -92,8 +91,7 @@ export function deleteFile(path: string): Promise<Result<string, Error>> {
       }
 
       return resolve(ok(`Path ${path} was not a file`))
-    }
-    catch (error) {
+    } catch (error) {
       return reject(err(error))
     }
   })
@@ -119,14 +117,15 @@ export async function deleteGlob(path: string): Promise<Result<string, Error>> {
 }
 
 export async function del(path: string): Promise<Result<string, Error>> {
-  if (isFile(path))
-    return await deleteFile(path)
+  if (isFile(path)) return await deleteFile(path)
 
-  if (isFolder(path))
-    return await deleteFolder(path)
+  if (isFolder(path)) return await deleteFolder(path)
 
-  if (path.includes('*'))
-    return await deleteGlob(path)
+  if (path.includes('*')) return await deleteGlob(path)
 
-  return err(handleError(`Path ${path} cannot be deleted due to an unhandled condition. Please report this issue.`))
+  return err(
+    handleError(
+      `Path ${path} cannot be deleted due to an unhandled condition. Please report this issue.`,
+    ),
+  )
 }
